refactor: move RPC conversions out of domain crates
Move protobuf/RPC-specific conversion logic from network and secrets into
the rpc crate, so crates that only need network or secrets types do not
also pull in RPC dependencies.

This keeps domain crates lighter and avoids forcing RPC/protobuf coupling
onto consumers that only need network/security functionality.
