+++ crates/network/Cargo.toml
@@ -49,8 +49,6 @@ sqlx = { workspace = true, features = [
 ], optional = true }
 thiserror = { workspace = true }
 
-carbide-rpc = { path = "../rpc" }
-
 [dev-dependencies]
 serde_json = { workspace = true }
