+++ crates/rpc/src/lib.rs
@@ -62,7 +62,9 @@ pub use crate::protos::{fmds, health, site_explorer};
 
 pub mod errors;
 pub mod forge_tls_client;
+pub mod network;
 pub mod protos;
+pub mod secrets;
 
 #[cfg(feature = "cli")]
 pub mod admin_cli;
