+++ crates/secrets/Cargo.toml
@@ -52,7 +52,6 @@ vaultrs = { workspace = true }
 
 # [local-dependencies]
 bmc-vendor = { path = "../bmc-vendor" }
-carbide-rpc = { path = "../rpc" }
 carbide-uuid = { path = "../uuid" }
 
 [dev-dependencies]
